+++ R/cpp11.R
@@ -12,6 +12,10 @@ dist_l1 <- function(x, y) {
   .Call(`_bases_dist_l1`, x, y)
 }
 
+row_prod <- function(x) {
+  .Call(`_bases_row_prod`, x)
+}
+
 im2col <- function(x, h, w, c, size, stride) {
   .Call(`_bases_im2col`, x, h, w, c, size, stride)
 }