+++ src/cpp11.cpp
@@ -26,6 +26,13 @@ extern "C" SEXP _bases_dist_l1(SEXP x, SEXP y) {
     return cpp11::as_sexp(dist_l1(cpp11::as_cpp<cpp11::decay_t<const doubles_matrix<>>>(x), cpp11::as_cpp<cpp11::decay_t<const doubles_matrix<>>>(y)));
   END_CPP11
 }
+// dist.cpp
+doubles row_prod(const doubles_matrix<> x);
+extern "C" SEXP _bases_row_prod(SEXP x) {
+  BEGIN_CPP11
+    return cpp11::as_sexp(row_prod(cpp11::as_cpp<cpp11::decay_t<const doubles_matrix<>>>(x)));
+  END_CPP11
+}
 // im2col.cpp
 doubles_matrix<> im2col(const doubles& x, int h, int w, int c, int size, int stride);
 extern "C" SEXP _bases_im2col(SEXP x, SEXP h, SEXP w, SEXP c, SEXP size, SEXP stride) {
@@ -40,6 +47,7 @@ static const R_CallMethodDef CallEntries[] = {
     {"_bases_dist_l2",    (DL_FUNC) &_bases_dist_l2,    2},
     {"_bases_forest_mat", (DL_FUNC) &_bases_forest_mat, 4},
     {"_bases_im2col",     (DL_FUNC) &_bases_im2col,     6},
+    {"_bases_row_prod",   (DL_FUNC) &_bases_row_prod,   1},
     {NULL, NULL, 0}
 };
 }